refactor: header and headeButton inculding UserAvatar and LoginRegister
