+++ src/components/header/Header.js
@@ -1,19 +1,25 @@
 import "../../style/header.scss";
-import Login, { UserAvatar } from "./UserAvatar";
+import { UserAvatar } from "./UserAvatar";
+import { LoginRegister } from "./LoginRegister"
 import { useSelector } from "react-redux";
 
 
 export function Header() {
   const loginInfo = useSelector((state) => state.user.value.loginInfo);
-  if (loginInfo.login === true) {
-    return(
-      <UserAvatar />
-    )
-  } else if (loginInfo.login === false) {
-    return (
-      <Login />
-    );
+  let headerButton;
+  if (loginInfo.login === false){
+    headerButton = <LoginRegister />
+  } else if (loginInfo.login === true) {
+    headerButton = <UserAvatar />
   }
+  return (
+    <header>
+      <div className="tool-bar">
+        <div className="logo">RealD</div>
+          {headerButton}
+      </div>
+    </header>
+  )
 }
 
 export default Header;
